requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
python-dotenv==1.0.1
linkedin-api==2.0.0a3
pandas==2.2.1
python-dateutil==2.8.2
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.4
pytz==2024.1
six==1.16.0
tqdm==4.66.2 